Se hace los arreglos para Regiones y  se agrega algunos modulos para Autenticación, existen problemas en la configuración reactiva pero se solucionará.

+++ src/main/java/com/bolsadeideas/springboot/backend/apirest/models/entity/Ids.java
@@ -20,14 +20,14 @@ public class Ids {
     private String name;
 
     @Column(value = "code")
-    private Long code;
+    private Integer code;
 
-    public Long getNewCode() {
+    public Integer getNewCode() {
         if (Optional.ofNullable(code).isPresent()) {
             return this.code + 1;
         } else {
             log.error("No se obtiene codigo");
-            return 0L;
+            return 0;
         }
     }
 }
